+++ lineage/figures/figureCommon.py
@@ -11,7 +11,7 @@ from matplotlib import gridspec, pylab, pyplot as plt
 import seaborn as sns
 import svgutils.transform as st
 from ..Analyze import run_Results_over, run_Analyze_over
-from ..LineageInputOutput import CensoredRecursive
+from ..plotTree import CensoredRecursive
 
 from ..states.StateDistributionGamma import StateDistribution
 from ..states.StateDistributionExpon import StateDistribution as expon_state
@@ -161,20 +161,6 @@ def overlayCartoon(figFile, cartoonFile, x, y, scalee=1, scale_x=1, scale_y=1):
 
     template.append(cartoon)
     template.save(figFile)
-
-def plotLineage(lineage):
-    """
-    Makes lineage tree.
-    """
-
-    a = [Clade(lineage.full_lineage[0].obs[2]+lineage.full_lineage[0].obs[3])]
-
-    # input the root cells in the lineage
-    c = CensoredRecursive(lineage.full_lineage[0], a)
-
-    Phylo.draw(c)
-    pylab.axis('off')
-    pylab.savefig('lineage/figures/cartoons/lineageTree.svg', format='svg', bbox_inches='tight', dpi=300)
 
 def figureMaker(ax, x, paramEst, accuracies, tr, pii, paramTrues, xlabel="Number of Cells"):
     """
